+++ src/components/sections/docsSection.vue
@@ -1,12 +1,12 @@
 <template>
   <div>
-    <article v-if="pattern.mdFile" id="markdown-body" class="markdown-body"
+    <article v-if="$store.state.pattern.sections.docs.sections.documentation.visible && pattern.mdFile" id="markdown-body" class="markdown-body"
              v-html="marked(pattern.mdFile.contents)">
     </article>
-    <div v-else class="notification is-warning">
+    <div v-if="$store.state.pattern.sections.docs.sections.documentation.visible && !pattern.mdFile" class="notification is-warning">
       There is no documentation available.
     </div>
-    <aside class="menu">
+    <aside v-if="$store.state.pattern.sections.docs.sections.includes.visible" class="menu">
       <p class="menu-label">Includes</p>
       <ul class="menu-list">
         <li v-for="(item, index) of twigIncludes" v-bind:key="index">
@@ -15,7 +15,7 @@
         <li v-if="twigIncludes.length <= 0" class="has-text-grey">This pattern does not include other patterns.</li>
       </ul>
     </aside>
-    <div class="content margin-top">
+    <div v-if="$store.state.pattern.sections.docs.sections.assets.visible" class="content margin-top">
       <p class="is-uppercase has-text-grey is-size-7">Assets</p>
       <div v-if="twigFile">
         <button class="button is-tag is-small is-warning" @click="copy(twigFile.rawPath.replace(patternPath, ''))">
@@ -36,8 +36,8 @@
         {{ file.rawPath.replace(patternPath, '') }}
       </div>
 
-      <div v-if="pattern.jsFiles.length <= 0 && pattern.cssFiles.length <= 0" class="notification is-warning">
-        There are no assets for this pattern.
+      <div v-if="$store.state.pattern.sections.docs.sections.assets.visible && (pattern.jsFiles.length <= 0 && pattern.cssFiles.length <= 0)" class="notification is-warning">
+        There are no CSS or JS assets for this pattern.
       </div>
     </div>
   </div>
@@ -63,11 +63,11 @@
         // if there are no include statements then return an empty array
         if (!this.includeStatements || this.includeStatements.length === 0) {
           return []
-        }        
+        }
 
         let statements = this.includeStatements.map((statement) => {
           // statement example: "include 'atoms/decorators/shape/shape.twig"
-          
+
           let name = statement.substring(
             statement.lastIndexOf('/') + 1,
             statement.lastIndexOf('.twig')
@@ -147,7 +147,7 @@
 
         includeStatements.forEach(statement => {
           // statement example: "include 'atoms/decorators/shape/shape.twig"
- 
+
           if (statement.startsWith('include \'/')) {
             errorMsg.push({
               type: 'include',
@@ -159,7 +159,7 @@
               type: 'include',
               message: 'For consistency purposes twig includes should use \' instead of \": ' + statement + '. '
             })
-          }          
+          }
         })
 
         this.$eventHub.$emit('includeErrors', errorMsg)